+++ test/cookbooks/selinux_test/recipes/enforcing.rb
@@ -1,4 +1,4 @@
-if platform?('ubuntu')
+if platform_family?('debian')
   # Ubuntu breaks kitchen SSH connections by default so need to load a module or two first
   include_recipe 'selinux::permissive'
 else