+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 This file is used to list changes made in each version of the selinux cookbook.
 
+## Unreleased
+
 ## 6.0.5 - *2022-09-18*
 
 - Standardise files with files in sous-chefs/repo-management